backend/requirements.txt
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Authlib==1.6.0
bcrypt==4.0.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.8
fastapi-mail==1.5.0
google-auth==2.40.3
greenlet==3.1.1
h11>=0.16,<0.17
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.26.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
redis==6.2.0
requests==2.32.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
